requirements.txt
discord.py
Pillow
flask